--- tests/_test_projects/mobile/settings.json
@@ -1,1 +1,0 @@
-{"root": "C:\\Users\\Justi\\OneDrive\\Documents\\Projects\\Python\\SpreadCounter\\tests\\_test_projects\\mobile", "counts": {"0": {"count": 90, "count_offset": 0, "input_image": "20201006_192404.jpg"}, "1": {"count": 124, "count_offset": 0, "input_image": "20201006_192418.jpg"}, "2": {"count": 92, "count_offset": 0, "input_image": "20201006_192432.jpg"}, "3": {"count": 163, "count_offset": 0, "input_image": "20201006_192446.jpg"}, "4": {"count": 129, "count_offset": 0, "input_image": "20201006_192449.jpg"}}, "ui_settings": {"draw_centers_cb": true, "draw_circles_cb": true, "draw_dish_circles_cb": true, "draw_count_cb": false, "count_offset_sb": 0, "upper_thresh_sb": 255, "lower_thresh_sb": 200, "circle_min_dist_sb": 5, "circle_min_rad_sb": 1, "circle_max_rad_sb": 10, "image_scale_factor_sb": 0.2, "contrast_multiplier_sb": 1.2, "max_dish_offset_radius_sb": 0.5, "min_dish_offset_radius_sb": 0.2, "dish_offset_sb": 0.85, "thresholding_cb": 0, "dish_detection_gb": true}}
